pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "qibo"
version = "0.2.19"
description = "A framework for quantum computing with hardware acceleration."
authors = ["The Qibo team"]
license = "Apache License 2.0"
readme = "README.md"
homepage = "https://qibo.science/"
repository = "https://github.com/qiboteam/qibo/"
documentation = "https://qibo.science/docs/qibo/stable"
keywords = []
classifiers = [
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering :: Physics",
]
packages = [{ include = "qibo", from = "src" }]
include = ["*.out", "*.yml"]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
numpy = "^2.0.0"
scipy = "^1.15.2"
sympy = "^1.13.1"
networkx = "^3.2.1"
cma = "^3.4.0"
joblib = "^1.4.2"
optuna = "^4.2.1"
tabulate = "^0.9.0"
openqasm3 = { version = "^1.0.1", extras = ["parser"] }
qulacs = { version = "^0.6.4", optional = true, python = "<3.13" }
perceval-quandela = ">=0.13"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pdbpp = "^0.10.3"
ipython = "^8.34.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^6.1.3"
furo = "^2022.12.7"
recommonmark = "^0.7.1"
sphinxcontrib-bibtex = "^2.5.0"
sphinx-markdown-tables = "^0.0.17"
sphinx-copybutton = "^0.5.2"
nbsphinx = "^0.8.12"
qulacs = { version = "^0.6.4", python = "<3.13" }
seaborn = "^0.13.2"
ipykernel = "^6.29.4"
qibojit = { git = "https://github.com/qiboteam/qibojit.git" }

[tool.poetry.group.tests]
optional = true

[tool.poetry.group.tests.dependencies]
pytest = "^7.2.1"
# cirq = "^1.4.1"
pytest-cov = "^4.0.0"
pylint = "^3.3.5"
matplotlib = "^3.7.0"
qibojit = { git = "https://github.com/qiboteam/qibojit.git" }
# qibotn = { git = "https://github.com/qiboteam/qibotn.git" }
# qiboml = { git  = "https://github.com/qiboteam/qiboml.git", branch="numpy_dep" }
stim = "^1.12.0"
qulacs = { version = "^0.6.4", python = "<3.13" }

[tool.poe.tasks]
test = "pytest"
lint = "pylint src --errors-only"
lint-warnings = "pylint src --exit-zero"
docs = "make -C doc html"
docs-clean = "make -C doc clean"
test-docs = "make -C doc doctest"


[tool.poetry.group.cuda11]
optional = true

[tool.poetry.group.cuda11.dependencies]
cupy-cuda11x = "^13.1.0"
cuquantum-python-cu11 = "^24.1.0"
qibojit = { git = "https://github.com/qiboteam/qibojit.git" }
# qibotn = { git = "https://github.com/qiboteam/qibotn.git" }


[tool.poetry.group.cuda12]
optional = true

[tool.poetry.group.cuda12.dependencies]
cupy-cuda12x = "^13.1.0"
cuquantum-python-cu12 = "^24.1.0"
qibojit = { git = "https://github.com/qiboteam/qibojit.git" }
# qibotn = { git = "https://github.com/qiboteam/qibotn.git" }

[tool.poetry.extras]
qinfo = ["stim"]
qulacs = ["qulacs"]

[tool.pylint.main]
ignored-modules = ["cvxpy", "torch"]

[tool.pylint.reports]
output-format = "colorized"

[tool.pytest.ini_options]
testpaths = ['tests/']
filterwarnings = ['ignore::RuntimeWarning']
addopts = [
  '--cov=qibo',
  '--cov-append',
  '--cov-report=xml',
  '--cov-report=html',
  '--durations=60',
]
